Don't crash when trying to delete filter that doesn't exist
The error can be repro'd by using "Add Filter" in a review,
then keeping it open in 2 tabs and deleting the added filter
in both tabs.

+++ src/operation/manipulatefilters.py
@@ -403,6 +403,12 @@ class RemoveReviewFilter(Operation):
 
         cursor.execute("SELECT review FROM reviewfilters WHERE id=%s", (filter_id,))
         review_id = cursor.fetchone()
+        if not review_id:
+            raise OperationFailure(
+                code="nosuchfilter",
+                title="No such filter!",
+                message=("Maybe the filter has been deleted since you "
+                         "loaded this page?"))
 
         cursor.execute("DELETE FROM reviewfilters WHERE id=%s", (filter_id,))
